Quynh1

+++ WebApplication2/Areas/Admin/Controllers/userAdminController.cs
@@ -1,10 +1,16 @@
 ﻿using DoAnCoSoAPI.Entities;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using MongoDB.Driver;
+using System.Security.Claims;
+using System.Security.Cryptography;
+using System.Text;
+using WebApplication2.Models;
 
 namespace WebApplication2.Areas.Admin.Controllers
 {
     [Area("admin")]
+
     public class userAdminController : Controller
     {
         private readonly IMongoCollection<User> _userCollection;
@@ -22,13 +28,20 @@ namespace WebApplication2.Areas.Admin.Controllers
         [HttpPost("Admin/userAdmin/Register")]
         public async Task<IActionResult> Register(User user)
         {
+            var userId = HttpContext.User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            var user1 = await _userCollection.Find(u => u.Id == userId).FirstOrDefaultAsync();
+
+            if (user1 == null || user1.role != "admin")
+            {
+                return BadRequest("Không phải Admin");
+            }
             if (user == null)
             {
                 return BadRequest("Invalid user object");
             }
 
             // Check if user with the same email already exists
-            var existingUser = await _userCollection.Find(u => u.Email == user.Email).FirstOrDefaultAsync();
+            var existingUser = await _userCollection.Find(u => u.eMail == user.eMail).FirstOrDefaultAsync();
             if (existingUser != null)
             {
                 return Conflict("User with this email already exists");
@@ -39,22 +52,54 @@ namespace WebApplication2.Areas.Admin.Controllers
 
             // You should hash the password before saving it to the database
             // For demonstration purposes, let's assume PasswordHash is already hashed
-            // user.PasswordHash = HashFunction(user.PasswordHash);
+            user.PasswordHash = HashPassword(user.PasswordHash);
 
             await _userCollection.InsertOneAsync(user);
             return Redirect("/admin/userAdmin/Index");
         }
-        [HttpGet]
-        public IActionResult Index()
+        public static string HashPassword(string password)
         {
-            // Retrieve list of users from the database and pass it to the view
-            var users = _userCollection.Find(u => true).ToList();
+            using (SHA256 sha256Hash = SHA256.Create())
+            {
+                // Băm mật khẩu
+                byte[] bytes = sha256Hash.ComputeHash(Encoding.UTF8.GetBytes(password));
+
+                // Chuyển đổi byte array sang chuỗi hex
+                StringBuilder builder = new StringBuilder();
+                for (int i = 0; i < bytes.Length; i++)
+                {
+                    builder.Append(bytes[i].ToString("x2"));
+                }
+                return builder.ToString();
+            }
+        }
+        [HttpGet]
+        public async Task<IActionResult> Index()
+        {
+            var userId = HttpContext.User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            var user = await _userCollection.Find(u => u.Id == userId).FirstOrDefaultAsync();
+
+            if (user == null || user.role != "admin")
+            {
+                return BadRequest("Không phải Admin");
+            }
+
+            var users = await _userCollection.Find(u => true).ToListAsync();
             return View(users);
         }
+
+
         [HttpGet("Admin/userAdmin/Edit/{id}")]
         public async Task<IActionResult> Edit(string id)
         {
             // Tìm người dùng theo id
+            var userId = HttpContext.User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            var admin = await _userCollection.Find(u => u.Id == userId).FirstOrDefaultAsync();
+
+            if (admin == null || admin.role != "admin")
+            {
+                return BadRequest("Không phải Admin");
+            }
             var user = await _userCollection.Find(u => u.Id == id).FirstOrDefaultAsync();
             if (user == null)
             {
@@ -71,8 +116,16 @@ namespace WebApplication2.Areas.Admin.Controllers
                 return BadRequest();
             }
 
+            // Kiểm tra xem người dùng đã cung cấp mật khẩu mới hay không
+            if (!string.IsNullOrEmpty(user.PasswordHash))
+            {
+                // Băm mật khẩu mới
+                user.PasswordHash = HashPassword(user.PasswordHash);
+            }
+
             // Cập nhật thông tin người dùng
             var result = await _userCollection.ReplaceOneAsync(u => u.Id == id, user);
+
             if (result.IsAcknowledged && result.ModifiedCount > 0)
             {
                 return RedirectToAction("Index");
@@ -83,10 +136,18 @@ namespace WebApplication2.Areas.Admin.Controllers
             }
         }
 
+
         [HttpGet("Admin/userAdmin/Delete/{id}")]
         public async Task<IActionResult> Delete(string id)
         {
             // Tìm người dùng theo id và hiển thị trang xác nhận xóa
+            var userId = HttpContext.User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            var admin = await _userCollection.Find(u => u.Id == userId).FirstOrDefaultAsync();
+
+            if (admin == null || admin.role != "admin")
+            {
+                return BadRequest("Không phải Admin");
+            }
             var user = await _userCollection.Find(u => u.Id == id).FirstOrDefaultAsync();
             if (user == null)
             {
@@ -97,7 +158,13 @@ namespace WebApplication2.Areas.Admin.Controllers
 
         [HttpPost("Admin/userAdmin/Delete/{id}")]
         public async Task<IActionResult> DeleteConfirmed(string id)
-        {
+        {var userId = HttpContext.User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            var admin = await _userCollection.Find(u => u.Id == userId).FirstOrDefaultAsync();
+
+            if (admin == null || admin.role != "admin")
+            {
+                return BadRequest("Không phải Admin");
+            }
             // Xóa người dùng từ database
             var result = await _userCollection.DeleteOneAsync(u => u.Id == id);
             if (result.IsAcknowledged && result.DeletedCount > 0)
@@ -123,7 +190,7 @@ namespace WebApplication2.Areas.Admin.Controllers
             }
 
             Console.WriteLine("Searching for users with email containing the provided value.");
-            var usersFound = await _userCollection.Find(u => u.Email.Contains(email)).ToListAsync();
+            var usersFound = await _userCollection.Find(u => u.eMail.Contains(email)).ToListAsync();
             Console.WriteLine($"Found {usersFound.Count} users.");
 
             // Trả về view với danh sách người dùng tìm thấy
